Refactor user management code
